Fix declaration of stpcpy

+++ common/utils.c
@@ -49,7 +49,7 @@
  * @return a pointer to the terminating `\0' character of @s1,
  * or NULL if @s1 or @s2 is NULL.
  */
-char *stpcpy(char *s1, int size, const char *s2)
+char *stpcpy(char *s1, size_t size, const char *s2)
 {
 	if (s1 == NULL || s2 == NULL)
 		return NULL;
